Expo 410 bug 1 pkg 02.5

+++ src/role/role.controller.ts
@@ -104,7 +104,7 @@ export class RoleController {
     if (!existsGroup) {
       const newRoleGroup = await this.tagGroupService.create({
         name: 'Roles',
-        color: '#11111',
+        color: '#666666',
         isExclusive: false,
       });
 
@@ -261,11 +261,23 @@ export class RoleController {
     description: translate('route.role.update.already-exists'),
     type: ErrorDto,
   })
+  @ApiConflictResponse({
+    description: translate('route.role.update.conflict-in-use'),
+    type: ErrorDto,
+  })
   @Patch('/:id')
   async update(
     @Param('id', new ExistingRecord('tag')) id: string,
     @Body() updateRoleDto: UpdateRoleDto,
   ): Promise<z.infer<typeof updateRoleResponseSchema>> {
+    const isUsed = await this.roleService.isRoleInUse(id);
+
+    if (isUsed) {
+      throw new ConflictException(
+        translate('route.role.update.conflict-in-use'),
+      );
+    }
+
     const existsRole = await this.roleService.existsRole(updateRoleDto.name);
 
     if (existsRole) {
@@ -283,10 +295,22 @@ export class RoleController {
     description: translate('route.role.delete.success'),
     type: DeleteRoleResponseDto,
   })
+  @ApiConflictResponse({
+    description: translate('route.role.delete.conflict-in-use'),
+    type: ErrorDto,
+  })
   @Delete('/:id')
   async delete(
     @Param('id', new ExistingRecord('tag')) id: string,
   ): Promise<z.infer<typeof deleteRoleResponseSchema>> {
+    const isUsed = await this.roleService.isRoleInUse(id);
+
+    if (isUsed) {
+      throw new ConflictException(
+        translate('route.role.delete.conflict-in-use'),
+      );
+    }
+
     return this.roleService.delete(id);
   }
 }
